jobs/pyproject.toml
[project]
name = "fishsignal-jobs"
version = "0.1.0"
description = "FishSignal offline data processing jobs"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "anthropic>=0.40.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]
